backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2>=2.9  # if using PostgreSQL, remove if using SQLite
django-cors-headers>=4.0